refactor(withdraw): model flow state
Refs #45
